Commit web.js to require()
