+++ src/serialize.js
@@ -1,0 +1,59 @@
+module.exports = serialize
+
+const convert = require('./convert')
+const Serializer = require('./serializer')
+
+/**
+ * Serialize to the appropriate format
+ * @@ Currently NQuads and JSON/LD are deal with extrelemently inefficiently
+ * through mutiple conversions.
+ */
+function serialize (target, kb, base, contentType, callback) {
+  var documentString = null
+  try {
+    var sz = Serializer(kb)
+    var newSts = kb.statementsMatching(undefined, undefined, undefined, target)
+    var n3String
+    sz.suggestNamespaces(kb.namespaces)
+    sz.setBase(base)
+    switch (contentType) {
+      case 'application/rdf+xml':
+        documentString = sz.statementsToXML(newSts)
+        return executeCallback(null, documentString)
+      case 'text/n3':
+      case 'application/n3': // Legacy
+        documentString = sz.statementsToN3(newSts)
+        return executeCallback(null, documentString)
+      case 'text/turtle':
+      case 'application/x-turtle': // Legacy
+        sz.setFlags('si') // Suppress = for sameAs and => for implies
+        documentString = sz.statementsToN3(newSts)
+        return executeCallback(null, documentString)
+      case 'application/ld+json':
+        n3String = sz.statementsToN3(newSts)
+        convert.convertToJson(n3String, callback)
+        break
+      case 'application/n-quads':
+      case 'application/nquads': // @@@ just outpout the quads? Does not work for collections
+        n3String = sz.statementsToN3(newSts)
+        documentString = convert.convertToNQuads(n3String, callback)
+        break
+      default:
+        throw new Error('Serialize: Content-type ' + contentType + ' not supported for data write.')
+    }
+  } catch (err) {
+    if (callback) {
+      return callback(err)
+    }
+    throw err // Don't hide problems from caller in sync mode
+  }
+
+  function executeCallback (err, result) {
+    if (callback) {
+      callback(err, result)
+      return
+    } else {
+      return result
+    }
+  }
+}
